+++ src/main/java/home/local/vtbtest/repository/FileRepository.java
@@ -1,7 +1,7 @@
 package home.local.vtbtest.repository;
 
-import home.local.vtbtest.model.File;
+import home.local.vtbtest.entity.File;
 import org.springframework.data.repository.CrudRepository;
 
-public interface FileRepository extends CrudRepository<File, Integer> {
+public interface FileRepository extends CrudRepository<File, Long> {
 }
